Allow the Github organization to be specified on the Kabanero object This allows the CLI to query the right information in Github, for the organization this instance is using.  I refactored the code in the landing page that programmatically creates a Deployment since the CLI now needs to do the same.
