+++ config/orchestrations/cli-services/0.1/kabanero-cli.yaml
@@ -1,10 +1,3 @@
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: kabanero-cli-role-config
-data:
-  groupsForTeam_test_roles_kabanero_io: "admin"
----
 apiVersion: apps/v1
 kind: Deployment
 metadata:
