+++ deploy/crds/kabanero_kabanero_crd.yaml
@@ -25,6 +25,31 @@ spec:
             type: object
           spec:
             properties:
+              appsodyOperator:
+                properties:
+                  enable:
+                    type: boolean
+                  image:
+                    type: string
+                  repository:
+                    type: string
+                  tag:
+                    type: string
+                  version:
+                    type: string
+                type: object
+              cliServices:
+                properties:
+                  image:
+                    type: string
+                  repository:
+                    type: string
+                  tag:
+                    type: string
+                  version:
+                    description: 'Future: Enable     bool   `json:"enable,omitempty"`'
+                    type: string
+                type: object
               collections:
                 properties:
                   repositories:
@@ -41,10 +66,14 @@ spec:
                       type: object
                     type: array
                 type: object
+              githubOrganization:
+                type: string
               kappnav:
                 properties:
                   enable:
                     type: boolean
+                type: object
+              landing:
                 type: object
               tekton:
                 properties:
@@ -75,8 +104,6 @@ spec:
                     items:
                       type: string
                     type: array
-                  ready:
-                    type: string
                 type: object
               kabaneroInstance:
                 description: Kabanero operator instance readiness status. The status
@@ -160,6 +187,32 @@ spec:
             type: object
           spec:
             properties:
+              appsodyOperator:
+                properties:
+                  enable:
+                    type: boolean
+                  image:
+                    type: string
+                  repository:
+                    type: string
+                  tag:
+                    type: string
+                  version:
+                    type: string
+                type: object
+              cliServices:
+                properties:
+                  enable:
+                    type: boolean
+                  image:
+                    type: string
+                  repository:
+                    type: string
+                  tag:
+                    type: string
+                  version:
+                    type: string
+                type: object
               collections:
                 properties:
                   repositories:
@@ -176,7 +229,9 @@ spec:
                       type: object
                     type: array
                 type: object
-              kappnav:
+              githubOrganization:
+                type: string
+              landing:
                 properties:
                   enable:
                     type: boolean
@@ -210,22 +265,10 @@ spec:
                     items:
                       type: string
                     type: array
-                  ready:
-                    type: string
                 type: object
               kabaneroInstance:
                 description: Kabanero operator instance readiness status. The status
                   is directly correlated to the availability of resources dependencies.
-                properties:
-                  errorMessage:
-                    type: string
-                  ready:
-                    type: string
-                  version:
-                    type: string
-                type: object
-              kappnav:
-                description: Kabanero Application Navigator instance readiness status.
                 properties:
                   errorMessage:
                     type: string
