+++ deploy/crds/kabanero_v1alpha1_kabanero_crd.yaml
@@ -70,6 +70,8 @@ spec:
                     type: object
                   type: array
               type: object
+            githubOrganization:
+              type: string
             kappnav:
               properties:
                 enable:
