+++ deploy/crds/kabanero_v1beta1_kabanero_crd.yaml
@@ -71,6 +71,8 @@ spec:
                     type: object
                   type: array
               type: object
+            githubOrganization:
+              type: string
             landing:
               properties:
                 enable:
