+++ pkg/apis/kabanero/v1beta1/zz_generated.openapi.go
@@ -207,6 +207,12 @@ func schema_pkg_apis_kabanero_v1beta1_KabaneroSpec(ref common.ReferenceCallback)
 							Format: "",
 						},
 					},
+					"githubOrganization": {
+						SchemaProps: spec.SchemaProps{
+							Type:   []string{"string"},
+							Format: "",
+						},
+					},
 					"collections": {
 						SchemaProps: spec.SchemaProps{
 							Ref: ref("github.com/kabanero-io/kabanero-operator/pkg/apis/kabanero/v1beta1.InstanceCollectionConfig"),
